+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 
 * Diffable datasources exceptions in Xcode 13 / iOS 15 with some internal restructuring.
 
+### Removed
+
+* Deprecated support for `UICollectionViewDiffableDataSourceReference`
+
 ## [8.2.0](https://github.com/DenTelezhkin/DTCollectionViewManager/releases/tag/8.2.0)
 
 ### Changed